feat(ui): split up `ViewerToolbar`
And fix disabled state for buttons
